styled text field fake modification
