+++ src/pages/Contract/Flow/GeneralInfo.tsx
@@ -10,7 +10,8 @@ import Button from '@mui/material/Button'
 import Divider from '@mui/material/Divider'
 import Stepper from '@mui/material/Stepper'
 import { styled } from '@mui/material/styles'
-import StepLabel from '@mui/material/StepLabel'
+import StepLabel from '@mui/material/StepLabel' 
+import TextField from '@mui/material/TextField'
 import Typography from '@mui/material/Typography'
 import CardContent from '@mui/material/CardContent'
 import MuiStep, { StepProps } from '@mui/material/Step'
@@ -39,8 +40,6 @@ import SelectAdherent from 'src/shared-components/custom-select/SelectAdherentLi
 import Select from 'react-select'
 import SelectIndividu from 'src/shared-components/custom-select/SelectIndividuals'
 import CustomTextField from 'src/SharedComponents/StyledTextField/StyledTextField '
-
-import CustomDataGrid from 'src/shared-components/data-grid/dataGrid'
 
 interface State {
   password: string
@@ -96,9 +95,9 @@ const Step = styled(MuiStep)<StepProps>(({ theme }) => ({
 const GeneralInfo = ({ popperPlacement }: { popperPlacement: ReactDatePickerProps['popperPlacement'] }) => {
   // ** States
   const [date, setDate] = useState<DateType>(new Date())
-  const [, setContract] = useState<string>('')
   const [google, setGoogle] = useState<string>('')
   const [twitter, setTwitter] = useState<string>('')
+  const [, setContract] = useState<string>('')
   const [facebook, setFacebook] = useState<string>('')
   const [linkedIn, setLinkedIn] = useState<string>('')
   const [activeStep, setActiveStep] = useState<number>(0)
@@ -120,8 +119,9 @@ const GeneralInfo = ({ popperPlacement }: { popperPlacement: ReactDatePickerProp
       toast.success('Form Submitted')
     }
   }
-  const handleReset = () => {   
+  const handleReset = () => {
     setGoogle('')
+
     setTwitter('')
     setContract('')
     setLinkedIn('')
@@ -164,7 +164,7 @@ const GeneralInfo = ({ popperPlacement }: { popperPlacement: ReactDatePickerProp
         return (
           <Fragment>
             <Grid item sm={3}>
-              <CustomTextField fullWidth placeholder={'anis'} label='anis' name='anis' correctValue='anis' control={undefined} helperMessage={undefined} field={undefined} />
+              <CustomTextField fullWidth placeholder={'anis'} label='anis' name='anis' correctValue='anis' />
             </Grid>
             <Grid item sm={3}>
               <StyledLabel>Type contrat </StyledLabel>
@@ -176,7 +176,8 @@ const GeneralInfo = ({ popperPlacement }: { popperPlacement: ReactDatePickerProp
                 placeholder={'Tapez le chiffres d affaires '}
                 label='Chiffres d affaires'
                 name='ChiffreDaffaire'
-                correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
+                correctValue=''
+              />
             </Grid>
             <Grid item sm={3}>
               <CustomTextField
@@ -184,7 +185,8 @@ const GeneralInfo = ({ popperPlacement }: { popperPlacement: ReactDatePickerProp
                 placeholder={'Factures transmises avec '}
                 label='Factures transmises avec'
                 name='Factures transmises avec'
-                correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
+                correctValue=''
+              />
             </Grid>
             <Grid item sm={3}>
               <StyledLabel> Status Contract </StyledLabel>
@@ -200,7 +202,8 @@ const GeneralInfo = ({ popperPlacement }: { popperPlacement: ReactDatePickerProp
                 placeholder={'Tapez le Dont Export '}
                 label='Dont Export'
                 name='Dont Export'
-                correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
+                correctValue=''
+              />
             </Grid>
 
             <Grid item sm={3}>
@@ -209,7 +212,8 @@ const GeneralInfo = ({ popperPlacement }: { popperPlacement: ReactDatePickerProp
                 placeholder={'Tapez le Nbr d acheteur prévus '}
                 label='Nbr d acheteur prévus'
                 name="Nbr d 'acheteur prévus"
-                correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
+                correctValue=''
+              />
             </Grid>
 
             <Grid item sm={3}>
@@ -233,7 +237,8 @@ const GeneralInfo = ({ popperPlacement }: { popperPlacement: ReactDatePickerProp
                 placeholder={'Tapez le Dont dom '}
                 label='Dont dom'
                 name='Dont dom'
-                correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
+                correctValue=''
+              />
             </Grid>
             <Grid item sm={3}>
               <CustomTextField
@@ -241,7 +246,8 @@ const GeneralInfo = ({ popperPlacement }: { popperPlacement: ReactDatePickerProp
                 placeholder={'Dont dom '}
                 label='Chiffres d affaires'
                 name='Dont dom'
-                correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
+                correctValue=''
+              />
             </Grid>
             <Grid item sm={3}>
               <StyledLabel> Date Signature </StyledLabel>
@@ -270,7 +276,8 @@ const GeneralInfo = ({ popperPlacement }: { popperPlacement: ReactDatePickerProp
                 placeholder={'Limite de financement '}
                 label='Limite de financement'
                 name='Limite de financement'
-                correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
+                correctValue=''
+              />
             </Grid>
             <Grid item sm={3}>
               <CustomTextField
@@ -278,7 +285,8 @@ const GeneralInfo = ({ popperPlacement }: { popperPlacement: ReactDatePickerProp
                 placeholder={"Nbr d'avoirs Prévus"}
                 label="Nbr d'avoirs Prévus"
                 name="Nbr d'avoirs Prévus"
-                correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
+                correctValue=''
+              />
             </Grid>
 
             <Grid item sm={3}>
@@ -297,7 +305,8 @@ const GeneralInfo = ({ popperPlacement }: { popperPlacement: ReactDatePickerProp
                 placeholder={'Délai moyen de réglement (jour) '}
                 label='Délai moyen de réglement (jour)'
                 name='Délai moyen de réglement (jour)'
-                correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
+                correctValue=''
+              />
             </Grid>
             <Grid item sm={3}>
               <CustomTextField
@@ -305,7 +314,8 @@ const GeneralInfo = ({ popperPlacement }: { popperPlacement: ReactDatePickerProp
                 placeholder={' Délai Max de règlement '}
                 label=' Délai Max de règlement'
                 name=' Délai Max de règlement'
-                correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
+                correctValue=''
+              />
             </Grid>
             <Grid item sm={3}></Grid>
             <Grid item sm={3}></Grid>
@@ -315,7 +325,8 @@ const GeneralInfo = ({ popperPlacement }: { popperPlacement: ReactDatePickerProp
                 placeholder={' Délai Max de règlement '}
                 label=' Délai Max de règlement'
                 name=' Délai Max de règlement'
-                correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
+                correctValue=''
+              />
             </Grid>
             <Grid item sm={3}>
               <StyledLabel> Nbr de remise Prévues </StyledLabel>
@@ -324,176 +335,62 @@ const GeneralInfo = ({ popperPlacement }: { popperPlacement: ReactDatePickerProp
                 placeholder={'Nbr de remise Prévues '}
                 label='Nbr de remise Prévues'
                 name='Nbr de remise Prévues'
-                correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
+                correctValue=''
+              />
             </Grid>
           </Fragment>
         )
       case 1:
         return (
           <Fragment key={step}>
-            <Grid item xs={12} sm={4}>
-              <CustomTextField
-                     fullWidth
-                     placeholder={'Délai moyen de réglement (jour) '}
-                     label='Délai moyen de réglement (jour)'
-                     name='Délai moyen de réglement (jour)'
-                     correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
-            </Grid>
-            <Grid item xs={12} sm={4}>
-              <CustomTextField
-                     fullWidth
-                     placeholder={'Délai moyen de réglement (jour) '}
-                     label='Délai moyen de réglement (jour)'
-                     name='Délai moyen de réglement (jour)'
-                     correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
-            </Grid>
-            <Grid item xs={12} sm={4}>
-              <CustomTextField
-                     fullWidth
-                     placeholder={'Délai moyen de réglement (jour) '}
-                     label='Délai moyen de réglement (jour)'
-                     name='Délai moyen de réglement (jour)'
-                     correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
-            </Grid>
-            <Grid item xs={12} sm={4}>
-              <CustomTextField
-                     fullWidth
-                     placeholder={'Délai moyen de réglement (jour) '}
-                     label='Délai moyen de réglement (jour)'
-                     name='Délai moyen de réglement (jour)'
-                     correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
-            </Grid>
-            <Grid item xs={12} sm={4}>
-              <CustomTextField
-                     fullWidth
-                     placeholder={'Délai moyen de réglement (jour) '}
-                     label='Délai moyen de réglement (jour)'
-                     name='Délai moyen de réglement (jour)'
-                     correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
-            </Grid>
-            <Grid item xs={12} sm={4}>
-              <CustomTextField
-                     fullWidth
-                     placeholder={'Délai moyen de réglement (jour) '}
-                     label='Délai moyen de réglement (jour)'
-                     name='Délai moyen de réglement (jour)'
-                     correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
-            </Grid>
-            <Grid item xs={12} sm={4}></Grid>
-            <Grid item xs={12} sm={4}>
-              <CustomTextField
-                     fullWidth
-                     placeholder={'Délai moyen de réglement (jour) '}
-                     label='Délai moyen de réglement (jour)'
-                     name='Délai moyen de réglement (jour)'
-                     correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
-            </Grid>
-            <Grid item xs={12} sm={4}>
-              {' '}
-              <CustomTextField
-                     fullWidth
-                     placeholder={'Délai moyen de réglement (jour) '}
-                     label='Délai moyen de réglement (jour)'
-                     name='Délai moyen de réglement (jour)'
-                     correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
-            </Grid>
-            <Grid item xs={12} sm={4}></Grid>
-            <Grid item xs={12} sm={4}></Grid>
-            <Grid item xs={12} sm={4}>
-              {' '}
-              <CustomTextField
-                     fullWidth
-                     placeholder={'Délai moyen de réglement (jour) '}
-                     label='Délai moyen de réglement (jour)'
-                     name='Délai moyen de réglement (jour)'
-                     correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
-            </Grid>
-            <Grid item xs={12} sm={4}></Grid>
-            <Grid item xs={12} sm={4}></Grid>
-            <Grid item xs={12} sm={4}>
-              {' '}
-              <CustomTextField
-                     fullWidth
-                     placeholder={'Délai moyen de réglement (jour) '}
-                     label='Délai moyen de réglement (jour)'
-                     name='Délai moyen de réglement (jour)'
-                     correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
-            </Grid>
-            <Grid item xs={12} sm={4}></Grid>
-            <Grid item xs={12} sm={4}></Grid>
-            <Grid item xs={12} sm={4}>
-              {' '}
-              <CustomTextField
-                     fullWidth
-                     placeholder={'Délai moyen de réglement (jour) '}
-                     label='Délai moyen de réglement (jour)'
-                     name='Délai moyen de réglement (jour)'
-                     correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
-            </Grid>
-            <Grid item xs={12} sm={4}></Grid>
-            <Grid item xs={12} sm={4}></Grid>
-            <Grid item xs={12} sm={4}>
-              {' '}
-              <CustomTextField
-                     fullWidth
-                     placeholder={'Délai moyen de réglement (jour) '}
-                     label='Délai moyen de réglement (jour)'
-                     name='Délai moyen de réglement (jour)'
-                     correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
-            </Grid>
-   
             <Grid item xs={12} sm={6}>
               <Select className='basic-single' classNamePrefix='select' isClearable={isClearable} name='color' />
             </Grid>
             <Grid item xs={12} sm={6}></Grid>
             <Grid item xs={12} sm={6}></Grid>
             <Grid item xs={12} sm={6}></Grid>
-       
           </Fragment>
         )
       case 2:
         return (
           <Fragment key={step}>
-
             <Grid item xs={12} sm={6}>
-              <StyledLabel>aa</StyledLabel>
-              <SelectIndividu labelText='Nom adherent' onSearch={handleSearch} />
-            </Grid>
-            <Grid item sm={4}>
-              <CustomTextField
-                   fullWidth
-                   placeholder={'Délai moyen de réglement (jour) '}
-                   label='Délai moyen de réglement (jour)'
-                   name='Délai moyen de réglement (jour)'
-                   correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
-            </Grid>
-            <Grid item sm={4}>
-              <CustomTextField
-                   fullWidth
-                   placeholder={'Délai moyen de réglement (jour) '}
-                   label='Délai moyen de réglement (jour)'
-                   name='Délai moyen de réglement (jour)'
-                   correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
-            </Grid>
-            <Grid item sm={4}>
-              <StyledLabel>aa</StyledLabel>
-              <StyledDatePicker
-                selected={date}
-                id='basic-input'
-                popperPlacement={popperPlacement}
-                onChange={(date: Date) => setDate(date)}
-                customInput={<CustomInput label='' />}
+              <TextField
+                fullWidth
+                label='Twitter'
+                value={twitter}
+                onChange={e => setTwitter(e.target.value)}
+                placeholder='https://twitter.com/carterLeonard'
               />
             </Grid>
-            <Grid item sm={4}>
-              <CustomTextField
-                     fullWidth
-                     placeholder={'Délai moyen de réglement (jour) '}
-                     label='Délai moyen de réglement (jour)'
-                     name='Délai moyen de réglement (jour)'
-                     correctValue='' control={undefined} helperMessage={undefined} field={undefined}              />
+            <Grid item xs={12} sm={6}>
+              <TextField
+                fullWidth
+                label='Facebook'
+                value={facebook}
+                onChange={e => setFacebook(e.target.value)}
+                placeholder='https://facebook.com/carterLeonard'
+              />
             </Grid>
-</Fragment>
+            <Grid item xs={12} sm={6}>
+              <TextField
+                fullWidth
+                label='Google+'
+                value={google}
+                onChange={e => setGoogle(e.target.value)}
+                placeholder='https://plus.google.com/carterLeonard'
+              />
+            </Grid>
+            <Grid item xs={12} sm={6}>
+              <TextField
+                fullWidth
+                label='LinkedIn'
+                value={linkedIn}
+                onChange={e => setLinkedIn(e.target.value)}
+                placeholder='https://linkedin.com/carterLeonard'
+              />
+            </Grid>
+          </Fragment>
         )
       default:
         return 'Unknown Step'
