+++ src/pages/Contract/Flow/index.tsx
@@ -2,39 +2,14 @@ import React from 'react'
 import { useForm, Controller } from 'react-hook-form'
 import CustomTextField from '../../../SharedComponents/StyledTextField/StyledTextField '
 
-function Index() {
-  const { control, handleSubmit } = useForm()
+import GeneralInfo from './GeneralInfo'
 
-  const onSubmit = (data: any) => {
-    console.log(data)
-  }
-
+function index() {
+  const popperPlacement = 'bottom';
+  
   return (
-    <form onSubmit={handleSubmit(onSubmit)}>
-      <Controller
-        name='fullName'
-        control={control}
-        rules={{ required: true, minLength: 7 }}
-        render={({ field, fieldState }) => (
-          <CustomTextField
-            field={undefined} correctValue={undefined} placeholder='fullname'
-
-
-
-            label='Full Name'
-            control={control}
-            {...field}
-            error={fieldState.invalid}
-
-            helperMessage='aaa'          />
-          
-        )}
-      />
-
-      {/* Ajoutez d'autres champs de formulaire ici si nécessaire */}
-      <button type='submit'>Submit</button>
-    </form>
+      <GeneralInfo popperPlacement={popperPlacement} />
   )
 }
 
-export default Index
+export default index
